changed from Float32 to Float64

+++ scripts/cmd_vel2motor.py
@@ -7,7 +7,7 @@ Created on Wed Mar 24 21:48:47 2021
 """
 
 import rospy
-from std_msgs.msg import Float64, Float32
+from std_msgs.msg import Float64, Float64
 from geometry_msgs.msg import Twist
 
 def callback(twist):
@@ -15,10 +15,10 @@ def callback(twist):
     right = twist.linear.x/1.5 + twist.angular.z/3.0
     global left_back, left_front, right_back, right_front 
 
-    left_msg = Float32()
+    left_msg = Float64()
     left_msg.data = left*12.0
     
-    right_msg = Float32()
+    right_msg = Float64()
     right_msg.data = right*12.0
     
     left_back.publish(left_msg)
@@ -33,10 +33,10 @@ def listener():
 
     
     global left_back, left_front, right_back, right_front 
-    left_back = rospy.Publisher('/left_back_motor/command', Float32, queue_size=10)
-    left_front = rospy.Publisher('/left_front_motor/command', Float32, queue_size=10)
-    right_back = rospy.Publisher('/right_back_motor/command', Float32, queue_size=10)
-    right_front = rospy.Publisher('/right_front_motor/command', Float32, queue_size=10)
+    left_back = rospy.Publisher('/left_back_motor/command', Float64, queue_size=10)
+    left_front = rospy.Publisher('/left_front_motor/command', Float64, queue_size=10)
+    right_back = rospy.Publisher('/right_back_motor/command', Float64, queue_size=10)
+    right_front = rospy.Publisher('/right_front_motor/command', Float64, queue_size=10)
 
     
     rospy.Subscriber("/cmd_vel", Twist, callback)
